feat: use lambda to provide error reporter in ParquetFileRecordFormat
- FileRecordFormat object and all it's fields need to be serializable
in order to construct the Flink job graph. Even though StatsDErrorReporter is
made serializable, it contains StatsDErrorReporter, which in turn
contains more fields which may not be serializable.Hence, in order to
mitigate job graph creation failures, we wrap the error reporter inside a
serializable lambda. This is a common idiom to make un-serializable
fields serializable in Java 8: https://stackoverflow.com/a/22808112

[https://github.com/odpf/dagger/issues/108]